Move Header.svelte back into Header/
I think this was created outside of Header/ by accident.
