+++ services/ui/src/js/components/Header/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Header";
+export { default } from "./Header.svelte";
